Refactored MoneyDisplay and Money name since I had commited a few typos. I also created two control classes the interface Command is in charge of acting as an extra layer in between the future classes that will implement it and ExchangeCommand, which will calculate the value of the money after the transaction from one currency to another.

+++ src/main/java/es/Miulpgc/software/model/Money.java
@@ -1,4 +1,4 @@
 package es.Miulpgc.software.model;
 
-public record Money(int amount, Currency currency) {
+public record Money(double amount, Currency currency) {
 }
